🚚 Rename: 쿼리 싱글톤 로직 파일 이동
Rename: 쿼리 싱글톤 로직 파일 이동
